Evaluate and sync firewall rules when parole state changes.
Earlier we used to just toggle the appstandby firewall chain.
This mostly works because even though the firewall chain is
disabled, we ensure the rules are up-to-date but if the
idle state of an app changes at the same time as the parole
state change, it is possible that NPMS will not receive the idle
state change callback. So, the firewall rule for that app
would be obsolete until it is re-evaluated on some other trigger.

Bug: 160855861
Test: atest ./hostsidetests/net/src/com/android/cts/net/HostsideRestrictBackgroundNetworkTests.java
Change-Id: I3f8d0d172ac415fbd133b32c6ec5952c2743b6f0