Unify dispatching timeout constants
The same constant is defined today with slightly different names in both
WindowMananagerService and ActivityTaskManagerService. To prevent
proliferation, make a constant in aidl and use it here.

Also, the two constants used to be in either nanoseconds or
milliseconds.

Make all constants in milliseconds because aidl does not support "const
long", and 5 seconds' worth of nanoseconds does not fit into an int.

Bug: 143459140
Test: interact with device after booting
Test: adb shell dumpsys input

Change-Id: Ia9b3f56d8eddcc6f2f0132bac9254655b180837f

Change-Id: If4d40b63c41f6fb54b99d167da8db1924945f3ef