WindowStateAnimator: Finish removing rotation animation path
There is only one clause left in the screenAnimation conditional
that differentiates it from the default computeShownFrameLocked. This
somewhat neublous comment about expanding windows out 1 pixel from their
center in order to fill holes. This code already doesn't do what it says
it does though, it just expands to the bottom/right, not out from the
center. That means if this path is still important (though removing it
doesn't seem to cause any problems) we can accomplish the same effect by
just adding a global scale to the animation.

Bug: 161937501
Test: Existing tests pass
Change-Id: I2ef686976db0b24356f7fea7a7fcfdf54195d13b