WindowState/WindowStateAnimator: Remove cropping
Cropping is all handled by higher levels of the hierarchy and
we don't need this code anymore. We can examine it bit by bit.

First we see lots of code cropping to screen boundaries, this will
be handled by DisplayContent/Area or the physical construction of
the screen.

In a few other cases we crop to the CompatFrame, this is the same
as the SurfaceSIze and so crop doesn't do anything.

In split-screen resize mode we crop to the display size, but this
doesm't do anything.

Cropping to the window frame wont have any effect since its the surface
frame before scaling, but we scale the crop rect after using it.

The only case we have to examine is calculateSystemDecorRect, but in new
inset mode, this only produces a crop for windows which don't draw
behind the nav bar. This is now limited to pre N (or earlier?) apps
and shouldn't have any effect.

Bug: 161937501
Test: Existing tests pass
Change-Id: I3c8edef0bb0d0482487184071902f366ed1f4ebb