Media - Handle cleanup of pending notification entries
It is possible for notifications to be canceled prior to inflation,
which means there will not be a callback when the notification is
removed. Utitlize the NotifCollectionListener to get this event.

Fixes: 166315453
Test: manual, using RCNs to generate a race condition as they
create/cancel notifications immediately

Change-Id: I56d3ab9ef46d86c8623f9bac9152e8a2b8341c01
(cherry picked from commit 7b6bf301e1d7b5cc4eb7b15b0a188993a5d4fa49)