camera2: Remove partial result nodes for error frames.
When there's an capture error reported for a frame, remove its nodes in
mPartialResults map node, since they'll not be used by clients of
FrameTracker. This should also reduce slow memory pressure build up
because of mPartialResult growth as and when there are error frames
which have some partial results as well.

Bug: 167944895

Test: GCA on Pixel2, constant mode changes don't show steady state increase in
      mPartialResults map size (Basic validity)

Change-Id: I6de585deb24039321310ddbd5dccd9119b25b23d
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>