Changed logic about how to check if VRI replaced its surface.
SurfaceView checks if VRI hasn't changed its surface when deciding
whether to call updateRelativeZ in updateSurface. This is to give
surfaceReplaced a chance to send the updateRelativeZ in the same
transaction as swapping the surfaces.

However, SV was only checking generation id, which may not have changed
when using a blast layer. Instead, VRI will increment the sequence id
when the surface has been replaced. SV can check that value and only
do an immediate updateRelativeZ when the sequence id has not changed.

Test: Blast + Split + SV with no flicker
Change-Id: Icdf9f1fbe12fe77052030aa37d808517c5a6cd38