Reland "Distinguish warm/hot launch by whether the activity has attached process"
If the process of an activity was died in background, its activity and
task record will still be in history stack. If the process is started
again in background such as receiving broadcast. Then the launch of the
same activity will have the conditions: process exists and the result
code is START_TASK_TO_FRONT. But the activity still need to create so
it should be warm launch.

This reverts commit 9a2d2b11f243f68f9ab8e2e67b2ef811a80b3483.
Reason for revert: remove redundant "notifyWindowsDrawn(mTopActivity)"
in test case that was missed when resolving conflict.

Bug: 171549711
Test: ActivityMetricsLaunchObserverTests#testOnActivityLaunchFinished
Change-Id: I43625184392c22e6769a147beda4817f70afcf9a
Merged-In: I2b8e45ddfd76ce4fb0fdb0c244a8f4170fbabd21